refactor: centralize card prompt text normalization
