+++ packages/server/src/services/prompt/card-text.ts
@@ -1,0 +1,5 @@
+import { stripMacroComments } from "@marinara-engine/shared";
+
+export function cardPromptText(value: unknown): string {
+  return typeof value === "string" ? stripMacroComments(value).trim() : "";
+}
